Refactor SinglePixelRegion and SpectrumPerSpaxel to share code
